utils/error.c: add file and move ft_perror() functions
