+++ src/utils/command.c
@@ -1,4 +1,4 @@
-/******************************************************************************/
+/* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   command.c                                          :+:      :+:    :+:   */
@@ -6,9 +6,9 @@
 /*   By: nefimov <nefimov@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 19:24:15 by nefimov           #+#    #+#             */
-/*   Updated: 2025/07/01 22:06:44 by nefimov          ###   ########.fr       */
+/*   Updated: 2025/07/02 23:54:26 by nefimov          ###   ########.fr       */
 /*                                                                            */
-/******************************************************************************/
+/* ************************************************************************** */
 
 #include "execution.h"
 #include "minishell.h"
@@ -159,34 +159,4 @@ int	ft_close_all_fd(t_shell *shell)
 		cmd = cmd->next;
 	}
 	return (0);
-}
-
-int	ft_perror(char *process, char *arg, char *message, int code)
-{
-	if (process)
-	{
-		ft_putstr_fd(process, STDERR_FILENO);
-		ft_putstr_fd(": ", STDERR_FILENO);
-	}
-	if (arg)
-	{
-		ft_putstr_fd(arg, STDERR_FILENO);
-		ft_putstr_fd(": ", STDERR_FILENO);
-	}
-	if (message)
-	{
-		ft_putstr_fd(message, STDERR_FILENO);
-	}	
-	ft_putstr_fd("\n", STDERR_FILENO);
-	return (code);
-}
-
-int	ft_perror_syntax(char *arg)
-{
-	ft_putstr_fd("minishell: ", STDERR_FILENO);
-	ft_putstr_fd("syntax error near unexpected token `", STDERR_FILENO);
-	if (arg)
-		ft_putstr_fd(arg, STDERR_FILENO);
-	ft_putstr_fd("'\n", STDERR_FILENO);
-	return (2);
 }
