+++ src/utils/error.c
@@ -1,0 +1,45 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   error.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nefimov <nefimov@student.42berlin.de>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/02 23:53:40 by nefimov           #+#    #+#             */
+/*   Updated: 2025/07/02 23:55:10 by nefimov          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "execution.h"
+#include "minishell.h"
+#include "parser.h"
+
+int	ft_perror(char *process, char *arg, char *message, int code)
+{
+	if (process)
+	{
+		ft_putstr_fd(process, STDERR_FILENO);
+		ft_putstr_fd(": ", STDERR_FILENO);
+	}
+	if (arg)
+	{
+		ft_putstr_fd(arg, STDERR_FILENO);
+		ft_putstr_fd(": ", STDERR_FILENO);
+	}
+	if (message)
+	{
+		ft_putstr_fd(message, STDERR_FILENO);
+	}
+	ft_putstr_fd("\n", STDERR_FILENO);
+	return (code);
+}
+
+int	ft_perror_syntax(char *arg)
+{
+	ft_putstr_fd("minishell: ", STDERR_FILENO);
+	ft_putstr_fd("syntax error near unexpected token `", STDERR_FILENO);
+	if (arg)
+		ft_putstr_fd(arg, STDERR_FILENO);
+	ft_putstr_fd("'\n", STDERR_FILENO);
+	return (2);
+}
